Public api AddLivePullStreamInfoConfig.

+++ services/live/describe_live_certificate_detail.go
@@ -87,7 +87,7 @@ type DescribeLiveCertificateDetailResponse struct {
 	RequestId string `json:"RequestId" xml:"RequestId"`
 	Cert      string `json:"Cert" xml:"Cert"`
 	Key       string `json:"Key" xml:"Key"`
-	CertId    int    `json:"CertId" xml:"CertId"`
+	CertId    int64  `json:"CertId" xml:"CertId"`
 	CertName  string `json:"CertName" xml:"CertName"`
 }
