+++ services/live/describe_live_stream_online_user_num.go
@@ -89,7 +89,7 @@ type DescribeLiveStreamOnlineUserNumRequest struct {
 type DescribeLiveStreamOnlineUserNumResponse struct {
 	*responses.BaseResponse
 	RequestId       string         `json:"RequestId" xml:"RequestId"`
-	TotalUserNumber int            `json:"TotalUserNumber" xml:"TotalUserNumber"`
+	TotalUserNumber int64          `json:"TotalUserNumber" xml:"TotalUserNumber"`
 	OnlineUserInfo  OnlineUserInfo `json:"OnlineUserInfo" xml:"OnlineUserInfo"`
 }
