+++ services/live/describe_live_top_domains_by_flow.go
@@ -88,8 +88,8 @@ type DescribeLiveTopDomainsByFlowResponse struct {
 	RequestId         string     `json:"RequestId" xml:"RequestId"`
 	StartTime         string     `json:"StartTime" xml:"StartTime"`
 	EndTime           string     `json:"EndTime" xml:"EndTime"`
-	DomainCount       int        `json:"DomainCount" xml:"DomainCount"`
-	DomainOnlineCount int        `json:"DomainOnlineCount" xml:"DomainOnlineCount"`
+	DomainCount       int64      `json:"DomainCount" xml:"DomainCount"`
+	DomainOnlineCount int64      `json:"DomainOnlineCount" xml:"DomainOnlineCount"`
 	TopDomains        TopDomains `json:"TopDomains" xml:"TopDomains"`
 }
